Pakeista komentavimo galimybe, pridedas komentavimo langas. Irasant komentara jis neissisaugoja ir neatvaizduoja jo.

+++ demo-blogas/src/main/java/com/example/demoblogas/blog/BlogsService.java
@@ -22,7 +22,7 @@ public class BlogsService {
         return repository.save(blog);
     }
 
-    public Blog getBlogById(Long id) {
+    public Blog getBlogById(Integer id) {
 
         return repository.findById(id).get();
     }
@@ -37,7 +37,7 @@ public class BlogsService {
         return repository.findAll(pageRequest);
 
     }
-    public void deleteBlogById(Long id) {
+    public void deleteBlogById(Integer id) {
         repository.deleteById(id);
     }
